+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Fixed a strange error stack appearing when CTRL+C-ing a detached command in a CmdSet. This is an unresolved Deno issue, by the way.
 - Fixed build being behind upstream Deno runtime.
+- Fixed `fkcommit` not working as it should if ran from a directory nested away from the repository's root.
 
 ## [5.2.0] (10-01-2026)